refactor(eval): extract CLI entrypoint into eval_cli.py
Split eval.py (379→187 lines) by moving _build_parser, _build_backend,
_run_calibrate, _emit_jsonl, _find_project_root, and main() into the
new eval_cli.py (205 lines). Core eval logic (data classes, classify_case,
evaluate_rule, load_test_cases) stays in eval.py. Updated all test imports
and mock patch paths.
